+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * added `[` method for `date_xx` objects that preserves class attribute when
   subsetting
+* `%y+%` / `%y-%` can add/subtract years from date_xx objects
 
 # dint 1.0.0